Junos and VI model: support uint32 admin cost with preference add/subtract
Extend administrative cost (Juniper preference) from int to long to support the full uint32 range (0-4,294,967,295) required by Junos. Add support for `set policy-options policy-statement PS term TERM then preference add/subtract <p>`.

Changes:
- Update AbstractRoute._admin field from int to long with MAX_ADMIN_DISTANCE changed to 0xFFFFFFFFL
- Update all route subclasses, builders, and infrastructure to use long for admin cost
- Change SetAdministrativeCost to only accept LongExpr (removed IntExpr support)
- Add IncrementAdministrativeCost and DecrementAdministrativeCost LongExpr classes with clipping at 0 and MAX_ADMIN_DISTANCE
- Update Junos grammar to parse `then preference [add|subtract] <uint32>`
- Update PsThenPreference to support operators (ADD/SUBTRACT/SET) similar to PsThenLocalPreference
- Update ConfigurationBuilder extraction logic to handle preference operators
- Update TransferBDD BDD representation from 8 bits to 32 bits for admin cost
- Add TransferBDD support for Increment/Decrement operations
- Fix all route constructors, hashCode methods, test matchers, and test assertions throughout codebase
- Add test configuration for preference add/subtract parsing

Resolves https://github.com/batfish/batfish/issues/8808

---

Prompt:
```
Add support for Junos `set policy-options policy-statement PS term TERM then preference subtract <p>`. Refer to the Junos CLI reference in working/ -- there's probably an `add` as well?

We need to implement this fully: Juniper parsing and extraction, and vendor-independent model RoutingPolicy statement. We need to make sure that subtract and add are clipped to the valid range. And we need to support them in TransferBDD.
```

Further discussion:
- Expanded scope to change AbstractRoute._admin from int to long throughout entire VI model
- This eliminates need for dual IntExpr/LongExpr support and simplifies the implementation
- Updated BDD representation to use 32 bits instead of 8 bits for full uint32 range support

commit-id:d8a97815

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/EvpnType3Route.java
@@ -98,7 +98,7 @@ public final class EvpnType3Route extends EvpnRoute<EvpnType3Route.Builder, Evpn
 
   @JsonCreator
   private static EvpnType3Route jsonCreator(
-      @JsonProperty(PROP_ADMINISTRATIVE_COST) int admin,
+      @JsonProperty(PROP_ADMINISTRATIVE_COST) long admin,
       @JsonProperty(PROP_AS_PATH) @Nullable AsPath asPath,
       @JsonProperty(PROP_CLUSTER_LIST) @Nullable Set<Long> clusterList,
       @JsonProperty(PROP_COMMUNITIES) @Nullable CommunitySet communities,
